pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Yevgeniy Kap <kapustaevg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
loguru = "^0.7.2"
fastapi-cache2 = "^0.2.2"
redis = "^5.0.7"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
